[fix] make FormField how separate component
