+++ src/pages/messages/Messages.module.pcss
@@ -14,6 +14,8 @@
     flex-direction: column;
     padding: 0;
     width: 100%;
+    height: 77vh;
+    overflow: auto;
 }
 
 .li {
