--- src/pages/registration/components/FormField/FormField.module.pcss
@@ -1,35 +1,0 @@
-.placeholder-wrapper {
-    transition: .2s;
-    position:absolute;
-    left: 0;
-    
-    pointer-events: none; 
-    width: 100%;
-    transform: translate(0, -24px);
-    text-align: center;
-}
-
-.placeholder-wrapper span {
-    font-size: 14px;
-    color: grey;
-    width:fit-content;
-}
-
-.placeholder-wrapper-focused {
-    transition: .2s;
-    position:absolute;
-    left: 0;
-    
-    pointer-events: none; 
-    width: 100%;
-    transform: translate(-45px, -50px);
-    text-align: center;
-}
-
-.placeholder-wrapper-focused span {
-    padding: 0 5px;
-    border-radius: 10px;
-    font-size: 14px;
-    color: grey;
-    width:fit-content;
-}
