--- src/pages/registration/components/FormField/FormField.module.pcss.d.ts
@@ -1,5 +1,0 @@
-declare const styles: {
-  readonly placeholderWrapper: string;
-  readonly placeholderWrapperFocused: string;
-};
-export = styles;
