--- src/pages/registration/components/FormField/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './FormField';
